move ShortCitationTemplate and CrossReferenceTemplate into reference folder
