+++ src/main/cliapp/src/components/Templates/reference/ShortCitationTemplate.js
